move the stats context in the statistics folder
